Implementing dollar-handling in mid-rule actions
git-svn-id: https://bisoncpp.svn.sourceforge.net/svnroot/bisoncpp/trunk@312 264705d0-3b1c-0410-8509-e6d6d96411b3

+++ bisonc++/parser/handleatsign.cc
@@ -5,7 +5,7 @@
 // The @-return value is not specified in bison's documentation. Is it not
 // used? Using the @ feature implies lsp-needed, which is set by implication
 
-void Parser::handleAtSign(size_t pos, size_t nElements, Block &block) 
+void Parser::handleAtSign(size_t pos, int nElements, Block &block) 
 {
     int idx;                // extract the index, determine its length
     size_t replaceSize = 1 + extractIndex(&idx, pos + 1);
